refactored diffusion-dislocation
